+++ api/main.py
@@ -1,10 +1,11 @@
 import os
 
 from fastapi import FastAPI
+from fastapi.middleware.cors import CORSMiddleware
 
-from api.db import mongo
-from api.db.users_mongo import UsersMongo
-from api.user.user_controller import UserController
+import db.mongo as mongo
+from db.users_mongo import UsersMongo
+from user.user_controller import UserController
 from routers import example
 
 
@@ -20,6 +21,18 @@ example.user_controller = user_controller
 # Build app
 app = FastAPI()
 app.include_router(example.router)
+
+origins = [
+    "http://localhost:5173",
+]
+
+app.add_middleware(
+    CORSMiddleware,
+    allow_origins=origins,
+    allow_credentials=True,
+    allow_methods=["*"],
+    allow_headers=["*"],
+)
 
 
 # TODO remove later
