+++ api/requirements.txt
@@ -1,6 +1,5 @@
 fastapi==0.109.0
 uvicorn==0.27.0.post1
 motor==3.3.2
-pydantic==2.0.0
 python-jose[cryptography]==3.3.0
 python-multipart==0.0.7
