+++ api/user/user_controller.py
@@ -1,5 +1,5 @@
-from api.user.user import User
-from api.user.user_repo import UserRepo
+from user.user import User
+from user.user_repo import UserRepo
 
 
 class UserController:
